create object point
