+++ ptychodus/controller/probe/exposure.py
@@ -2,7 +2,7 @@ import logging
 
 from ...model.analysis import ExposureAnalyzer, ExposureMap
 from ...model.visualization import VisualizationEngine
-from ...view.object import ExposureDialog
+from ...view.probe import ExposureDialog
 from ...view.widgets import ExceptionDialog
 from ..data import FileDialogFactory
 from ..visualization import VisualizationParametersController, VisualizationWidgetController
